Configurações do jest e Mudanças gerais de boas práticas, obs:emprestimos ainda não está funcionando

+++ classes/Emprestimo.ts
@@ -3,37 +3,60 @@ import Prompt from "prompt-sync";
 import { Membro } from "./Membro";
 import { Livro } from "./Livro";
 
-export class Emprestimo {
-  protected _cpfMembro: string;
-  protected _ISBN_livro: string;
-  protected _dataEmprestimo: string;
-  protected _dataDevolucao: string;
+const prompt = Prompt();
 
-  constructor(cpfMembro: string, ISBN_livro: string, dataEmprestimo: string, dataDevolucao: string) {
-    this._cpfMembro = cpfMembro;
-    this._ISBN_livro = ISBN_livro;
+export class Emprestimo {
+  private _membro: Membro;
+  private _livro: Livro;
+  private _dataEmprestimo: string;
+  private _dataDevolucao: string;
+
+  constructor(membro: Membro, livro: Livro, dataEmprestimo: string, dataDevolucao: string) {
+    this._membro = membro;
+    this._livro = livro;
     this._dataEmprestimo = dataEmprestimo;
     this._dataDevolucao = dataDevolucao;
   }
 
+  private static carregarDados(filepath: string): any[] {
+    if (fs.existsSync(filepath)) {
+      const data = fs.readFileSync(filepath, "utf-8");
+      return JSON.parse(data);
+    } else {
+      return [];
+    }
+  }
+
   public adicionar(): void {
-''
+    const membros = Emprestimo.carregarDados("./data/membros.json");
+    const livros = Emprestimo.carregarDados("./data/livros.json");
+
+    const membroExiste = membros.some((membro: any) => membro._cpf === this._membro.cpf);
+    const livroExiste = livros.some((livro: any) => livro._isbn === this._livro.isbn);
+
+    if (!membroExiste) {
+      console.log("Membro não encontrado!");
+      return;
+    }
+
+    if (!livroExiste) {
+      console.log("Livro não encontrado!");
+      return;
+    }
 
     const emprestimoData = {
-      cpfMembro: this._cpfMembro,
-      ISBN_livro: this._ISBN_livro,
+      cpfMembro: this._membro.cpf,
+      ISBN_livro: this._livro.isbn,
       dataEmprestimo: this._dataEmprestimo,
       dataDevolucao: this._dataDevolucao,
     };
 
-    let emprestimos = [];
-    if (fs.existsSync("./data/emprestimos.json")) {
-      const data = fs.readFileSync("./data/emprestimos.json", "utf-8");
-      emprestimos = JSON.parse(data);
-    }
+    let emprestimos = Emprestimo.carregarDados("./data/emprestimos.json");
 
     emprestimos.push(emprestimoData);
     fs.writeFileSync("./data/emprestimos.json", JSON.stringify(emprestimos, null, 2));
+
+    console.log("Empréstimo adicionado com sucesso!");
   }
 
   public listar(): void {
@@ -48,15 +71,15 @@ export class Emprestimo {
   }
 
   public devolver(): void {
-    let key = Prompt();
-    let CPF = key("Digite o CPF do membro que deseja devolver o livro: ");
-    let ISBN = key("Digite o ISBN do livro que deseja devolver: ");
+    const CPF = prompt("Digite o CPF do membro que deseja devolver o livro: ");
+    const ISBN = prompt("Digite o ISBN do livro que deseja devolver: ");
 
     if (fs.existsSync("./data/emprestimos.json")) {
       const data = fs.readFileSync("./data/emprestimos.json", "utf-8");
       let emprestimos = JSON.parse(data);
+
       let emprestimo = emprestimos.find(
-        (e: Emprestimo) => e._cpfMembro === CPF && e._ISBN_livro === ISBN
+        (e: any) => e.cpfMembro === CPF && e.ISBN_livro === ISBN
       );
 
       if (emprestimo) {
@@ -72,79 +95,3 @@ export class Emprestimo {
     }
   }
 }
-
-
-
-
-
-// import fs from "fs";
-// import Prompt from "prompt-sync";
-// import { Membro } from "./Membro";
-// import { Livro } from "./Livro";
-
-// export class Emprestimo {
-//   protected _membro: Membro;
-//   protected _livro: Livro;
-//   protected _dataEmprestimo: string;
-//   protected _dataDevolucao: string;
-
-//   constructor(membro: Membro, livro: Livro, dataEmprestimo: string, dataDevolucao: string) {
-//     this._membro = membro;
-//     this._livro = livro;
-//     this._dataEmprestimo = dataEmprestimo;
-//     this._dataDevolucao = dataDevolucao;
-//   }
-
-//   public adicionar(): void {
-//     const emprestimoData = {
-//       membro: this._membro,
-//       livro: this._livro,
-//       dataEmprestimo: this._dataEmprestimo,
-//       dataDevolucao: this._dataDevolucao,
-//     };
-
-//     let emprestimos = [];
-//     if (fs.existsSync("./data/emprestimos.json")) {
-//       const data = fs.readFileSync("./data/emprestimos.json", "utf-8");
-//       emprestimos = JSON.parse(data);
-//     }
-
-//     emprestimos.push(emprestimoData);
-//     fs.writeFileSync("./data/emprestimos.json", JSON.stringify(emprestimos, null, 2));
-//   }
-
-//   public listar(): void {
-//     if (fs.existsSync("./data/emprestimos.json")) {
-//       const data = fs.readFileSync("./data/emprestimos.json", "utf-8");
-//       const emprestimos = JSON.parse(data);
-//       console.table(emprestimos);
-//     } else {
-//       console.log("Arquivo de empréstimos não encontrado!");
-//     }
-//   }
-
-//   public devolver(): void {
-//     let key = Prompt();
-//     let CPF = key("Digite o CPF do membro que deseja devolver o livro: ");
-//     let ISBN = key("Digite o ISBN do livro que deseja devolver: ");
-
-//     if (fs.existsSync("./data/emprestimos.json")) {
-//       const data = fs.readFileSync("./data/emprestimos.json", "utf-8");
-//       let emprestimos = JSON.parse(data);
-//       let emprestimo = emprestimos.find(
-//         (e: any) => e.membro.CPF === CPF && e.livro.ISBN === ISBN
-//       );
-
-//       if (emprestimo) {
-//         let index = emprestimos.indexOf(emprestimo);
-//         emprestimos.splice(index, 1);
-//         fs.writeFileSync("./data/emprestimos.json", JSON.stringify(emprestimos, null, 2));
-//         console.log("Livro devolvido com sucesso!");
-//       } else {
-//         console.log("Empréstimo não encontrado!");
-//       }
-//     } else {
-//       console.log("Arquivo de empréstimos não encontrado!");
-//     }
-//   }
-// }
